click one item on the table and go to detail UI
Feature 208

+++ Zuzu/NotificationItemsTableViewController.swift
@@ -14,7 +14,7 @@ struct NotificationItemsTableConst {
 
 class NotificationItemsTableViewController: UITableViewController {
 
-    var notificationItems = [NotificationHouseItem]()
+    var notificationItems: [NotificationHouseItem]?
     let notificationService = NotificationItemService.sharedInstance
     
     // MARK: - View Life Cycle
@@ -23,7 +23,7 @@ class NotificationItemsTableViewController: UITableViewController {
         super.viewDidLoad()
         doMockData()
         refresh()
-        NSLog("isRead:\(notificationItems[0].isRead)")
+        
         //tableView.estimatedRowHeight = tableView.rowHeight
         //tableView.rowHeight = UITableViewAutomaticDimension
         
@@ -46,7 +46,11 @@ class NotificationItemsTableViewController: UITableViewController {
     }
 
     override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return notificationItems.count
+        if let items = self.notificationItems{
+            return items.count
+        }
+        
+        return 0
     }
 
 
@@ -58,7 +62,7 @@ class NotificationItemsTableViewController: UITableViewController {
         let cell = tableView.dequeueReusableCellWithIdentifier(Storyboard.CellReuseIdentifier, forIndexPath: indexPath) as! NotificationItemTableViewCell
         
         // Configure the cell...
-        cell.notificationItem = notificationItems[indexPath.row]
+        cell.notificationItem = notificationItems![indexPath.row]
         return cell
     }
     
@@ -67,18 +71,28 @@ class NotificationItemsTableViewController: UITableViewController {
     func refresh(){
         if let result = self.notificationService.getAll(){
             self.notificationItems = result
-            NSLog("data count: \(notificationItems.count)")
-            self.tableView.reloadData()
+            NSLog("data count: \(notificationItems!.count)")
         }
-        tableView.reloadData()
+        self.tableView.reloadData()
     }
     
     func deleteRow(row: Int){
-        let item = notificationItems[row]
-        notificationItems.removeAtIndex(row)
-        self.notificationService.deleteItem(item)
+        if self.notificationItems != nil{
+            let item = notificationItems![row]
+            notificationItems!.removeAtIndex(row)
+            self.notificationService.deleteItem(item)
+        }
     }
-    
+
+    func setRead(item: NotificationHouseItem){
+        if item.isRead == true{
+            return
+        }
+        item.isRead = true
+        var updateData = Dictionary<String, AnyObject>()
+        updateData["isRead"] = true
+        self.notificationService.updateItem(item, dataToUpdate: updateData)
+    }
     
     // MARK: - swipe-left-to-delete
     
@@ -92,30 +106,95 @@ class NotificationItemsTableViewController: UITableViewController {
     // MARK: - select item action
     
     override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
+        
+        if let item = self.notificationItems?[indexPath.row]{
+            if item.isRead == false{
+                setRead(item)
+            }
 
-        let item = notificationItems[indexPath.row]
-        item.isRead = true
-        var data = Dictionary<String, AnyObject>()
-        data["isRead"] = item.isRead
-        self.notificationService.updateItem(item, dataToUpdate: data)
-
+            let selectedCell:UITableViewCell = tableView.cellForRowAtIndexPath(indexPath)!
+            selectedCell.contentView.backgroundColor = UIColor.whiteColor()
+            
+            let storyboard = UIStoryboard(name: "SearchStoryboard", bundle: nil)
+            let vc = storyboard.instantiateViewControllerWithIdentifier("HouseDetailView") as! HouseDetailViewController
+            let houseItem = item.toHouseItem()
+            vc.houseItem = houseItem
+            
+            //GA Tracker
+            self.trackEventForCurrentScreen(GAConst.Catrgory.Activity,
+                action: GAConst.Action.Activity.ViewItemPrice,
+                label: String(houseItem.price))
+            
+            self.trackEventForCurrentScreen(GAConst.Catrgory.Activity,
+                action: GAConst.Action.Activity.ViewItemSize,
+                label: String(houseItem.size))
+            
+            self.trackEventForCurrentScreen(GAConst.Catrgory.Activity,
+                action: GAConst.Action.Activity.ViewItemType,
+                label: String(houseItem.purposeType))
+            
+            self.showViewController(vc, sender: self)
+        }
     }
     
     func doMockData(){
         var data = Dictionary<String, AnyObject>()
-        data["source"] = 1
-        data["id"] = "12"
-        data["link"] = ""
-        data["mobile_link"] = ""
-        data["title"] = "中山馥臨捷運美宅"
-        data["title"] = "台北市中山區雙城街23巷"
+        data["source"] = 2
+        data["id"] = "aHR0cDovL3JlbnQuaG91c2VmdW4uY29tLnR3L3JlbnQvaG91c2UvODg1NDgwLw=="
+        data["link"] = "http://rent.housefun.com.tw/rent/house/885480/"
+        data["mobile_link"] = "http://rent.housefun.com.tw/mobile/rent/house/885480"
+        data["title"] = "稀有小帝寶精品電梯套房@可炊流理臺~耗資百萬裝潢˙"
+        data["city"] = 400
+        data["region"] = 404
+        data["purpose_type"] = 1
+        data["house_type"] = 1
+        data["price"] = 8300
+        data["size"] = 8
+        data["img"] = [
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A00120140000809173235a33480.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732351.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732352.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732353.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732364.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732365.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732366.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732367.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732368.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A001201400008091732369.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A0012014000080917323610.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A0012014000080917323611.jpg?t=2011",
+            "https://pic.hfcdn.com/res/DrawImage/ShowPic/600/0/HFRENT04/Z999/09d5524b-ce5c-44e4-8f0e-bf5be791e332/Z999A0012014000080917323612.jpg?t=2011"
+        ]
+        if self.notificationItems?.count < 1{
+            self.notificationService.addItem(data)
+        }
+        
+        
+        data = Dictionary<String, AnyObject>()
+        data["source"] = 2
+        data["id"] = "aHR0cDovL3JlbnQuNTkxLmNvbS50dy9yZW50LWRldGFpbC0zODI5NjU1Lmh0bWw="
+        data["link"] = "http://rent.591.com.tw/rent-detail-3829655.html"
+        data["mobile_link"] = "http://m.591.com.tw/mobile-detail.html?houseId=R3829655"
+        data["title"] = "近中山北路民權東路口透天二樓整層出租"
         data["city"] = 100
         data["region"] = 104
         data["purpose_type"] = 1
-        data["house_type"] = 1
-        data["price"] = 10000
-        data["size"] = 2000
-        self.notificationService.addItem(data)
+        data["house_type"] = 3
+        data["price"] = 15000
+        data["size"] = 22
+        data["img"] = [
+            "http://cp2.591.com.tw/house/active/2015/12/19/145053190841130908_600x600x579519.jpg",
+            "http://cp2.591.com.tw/house/active/2015/12/19/145053280315500503_600x600x579519.jpg",
+            "http://cp2.591.com.tw/house/active/2015/12/19/145053282033236000_600x600x579519.jpg",
+            "http://cp2.591.com.tw/house/active/2015/12/19/145053283321717807_600x600x579519.jpg",
+            "http://cp2.591.com.tw/house/active/2015/12/19/145053284171791107_600x600x579519.jpg",
+            "http://cp2.591.com.tw/house/active/2015/12/19/145053284620772901_600x600x579519.jpg",
+            "http://cp2.591.com.tw/house/active/2015/12/19/145053285239617402_600x600x579519.jpg"
+        ]
+        if self.notificationItems?.count < 2{
+            self.notificationService.addItem(data)
+        }
+        
     }
 
     /*
